# python-flask/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Ytel API version 3"
  version: "3.12"
  title: "Ytel API V3"
host: "api.ytel.com"
basePath: "/api/v3"
schemes:
- "https"
consumes:
- "application/x-www-form-urlencoded"
produces:
- "application/json"
paths:
  /template/view.json:
    post:
      tags:
      - "SharedShortCode"
      summary: "View Template"
      description: "View a Shared ShortCode Template"
      operationId: "template_view"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "TemplateId"
        in: "formData"
        description: "The unique identifier for a template object"
        required: true
        type: "string"
        format: "uuid"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.shared_short_code_controller"
...
  /recording/viewrecording.json:
    post:
      tags:
      - "Recording"
      summary: "View Recording"
      description: "Retrieve the recording of a call by its RecordingSid. This resource\
        \ will return information regarding the call such as start time, end time,\
        \ duration, and so forth."
      operationId: "recording_viewrecording"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "recordingsid"
        in: "formData"
        description: "The unique identifier for the recording."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.recording_controller"
  /recording/deleterecording.json:
    post:
      tags:
      - "Recording"
      summary: "Delete Recording"
      description: "Remove a recording from your Ytel account."
      operationId: "recording_deleterecording"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "recordingsid"
        in: "formData"
        description: "The unique identifier for a recording."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.recording_controller"
...
  /accounts/viewaccount.json:
    post:
      tags:
      - "Account"
      summary: "View Account"
      description: "Retrieve information regarding your Ytel account by a specific\
        \ date. The response object will contain data such as account status, balance,\
        \ and account usage totals."
      operationId: "accounts_viewaccount"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "Date"
        in: "formData"
        description: "Filter account information based on date."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.account_controller"
...
  /shortcode/getinboundsms.json:
    post:
      tags:
      - "SharedShortCode"
      summary: "List Inbound SMS"
      description: "List All Inbound ShortCode"
      operationId: "shortcode_getinboundsms"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "formData"
        description: "The page count to retrieve from the total results in the collection.\
          \ Page indexing starts at 1."
        required: false
        type: "integer"
        default: 1
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "pagesize"
        in: "formData"
        description: "Number of individual resources listed in the response per page"
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "from"
        in: "formData"
        description: "From Number to Inbound ShortCode"
        required: false
        type: "string"
      - name: "Shortcode"
        in: "formData"
        description: "Only list messages sent to this Short Code"
        required: false
        type: "string"
      - name: "Datecreated"
        in: "query"
        description: "Only list messages sent with the specified date"
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.shared_short_code_controller"
  /shortcode/sendsms.json:
    post:
      tags:
      - "SharedShortCode"
      summary: "Send SMS"
      description: "Send an SMS from a Ytel ShortCode"
      operationId: "shortcode_sendsms"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "shortcode"
        in: "formData"
        description: "The Short Code number that is the sender of this message"
        required: true
        type: "string"
      - name: "to"
        in: "formData"
        description: "A valid 10-digit number that should receive the message"
        required: true
        type: "string"
      - name: "templateid"
        in: "formData"
        description: "The unique identifier for the template used for the message"
        required: true
        type: "string"
        format: "uuid"
      - name: "data"
        in: "formData"
        description: "format of your data, example: {companyname}:test,{otpcode}:1234"
        required: true
        type: "string"
      - name: "Method"
        in: "formData"
        description: "Specifies the HTTP method used to request the required URL once\
          \ the Short Code message is sent."
        required: false
        type: "string"
        default: "GET"
      - name: "MessageStatusCallback"
        in: "formData"
        description: "URL that can be requested to receive notification when Short\
          \ Code message was sent."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.shared_short_code_controller"
...
  /recording/listrecording.json:
    post:
      tags:
      - "Recording"
      summary: "List Recordings"
      description: "Retrieve a list of recording objects."
      operationId: "recording_listrecording"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "formData"
        description: "The page count to retrieve from the total results in the collection.\
          \ Page indexing starts at 1."
        required: false
        type: "integer"
        default: 1
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "pagesize"
        in: "formData"
        description: "The count of objects to return per page."
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "Datecreated"
        in: "formData"
        description: "Filter results by creation date"
        required: false
        type: "string"
      - name: "callsid"
        in: "formData"
        description: "The unique identifier for a call."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.recording_controller"
...
  /template/lists.json:
    post:
      tags:
      - "SharedShortCode"
      summary: "List Templates"
      description: "List Shortcode Templates by Type"
      operationId: "template_lists"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "type"
        in: "formData"
        description: "The type (category) of template Valid values: marketing, authorization"
        required: false
        type: "string"
        default: "authorization"
      - name: "page"
        in: "formData"
        description: "The page count to retrieve from the total results in the collection.\
          \ Page indexing starts at 1."
        required: false
        type: "integer"
        default: 1
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "pagesize"
        in: "formData"
        description: "The count of objects to return per page."
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "Shortcode"
        in: "formData"
        description: "Only list templates of type"
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.shared_short_code_controller"
...
  /keyword/view.json:
    post:
      tags:
      - "SharedShortCode"
      summary: "View Keyword"
      description: "View a set of properties for a single keyword."
      operationId: "keyword_view"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "Keywordid"
        in: "formData"
        description: "The unique identifier of each keyword"
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.shared_short_code_controller"
  /keyword/lists.json:
    post:
      tags:
      - "SharedShortCode"
      summary: "List Keywords"
      description: "Retrieve a list of keywords associated with your Ytel account."
      operationId: "keyword_lists"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "formData"
        description: "The page count to retrieve from the total results in the collection.\
          \ Page indexing starts at 1."
        required: false
        type: "integer"
        default: 1
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "pagesize"
        in: "formData"
        description: "Number of individual resources listed in the response per page"
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "Keyword"
        in: "formData"
        description: "Only list keywords of keyword"
        required: false
        type: "string"
      - name: "Shortcode"
        in: "formData"
        description: "Only list keywords of shortcode"
        required: false
        type: "integer"
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.shared_short_code_controller"
  /shortcode/viewshortcode.json:
    post:
      tags:
      - "SharedShortCode"
      summary: "View Shortcode"
      description: "The response returned here contains all resource properties associated\
        \ with the given Shortcode."
      operationId: "shortcode_viewshortcode"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "Shortcode"
        in: "formData"
        description: "List of valid Shortcode to your Ytel account"
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.shared_short_code_controller"
  /shortcode/listshortcode.json:
    post:
      tags:
      - "SharedShortCode"
      summary: "List Shortcodes"
      description: "Retrieve a list of shortcode assignment associated with your Ytel\
        \ account."
      operationId: "shortcode_listshortcode"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "formData"
        description: "The page count to retrieve from the total results in the collection.\
          \ Page indexing starts at 1."
        required: false
        type: "integer"
        default: 1
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "pagesize"
        in: "formData"
        description: "Number of individual resources listed in the response per page"
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "Shortcode"
        in: "query"
        description: "Only list keywords of shortcode"
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.shared_short_code_controller"
  /shortcode/updateshortcode.json:
    post:
      tags:
      - "SharedShortCode"
      summary: "Update Shortcode"
      description: "Update Assignment"
      operationId: "shortcode_updateshortcode"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "Shortcode"
        in: "formData"
        description: "List of valid shortcode to your Ytel account"
        required: true
        type: "string"
      - name: "FriendlyName"
        in: "formData"
        description: "User generated name of the shortcode"
        required: false
        type: "string"
      - name: "CallbackUrl"
        in: "formData"
        description: "URL that can be requested to receive notification when call\
          \ has ended. A set of default parameters will be sent here once the call\
          \ is finished."
        required: false
        type: "string"
      - name: "CallbackMethod"
        in: "formData"
        description: "Specifies the HTTP method used to request the required StatusCallBackUrl\
          \ once call connects."
        required: false
        type: "string"
      - name: "FallbackUrl"
        in: "formData"
        description: "URL used if any errors occur during execution of InboundXML\
          \ or at initial request of the required Url provided with the POST."
        required: false
        type: "string"
      - name: "FallbackUrlMethod"
        in: "formData"
        description: "Specifies the HTTP method used to request the required FallbackUrl\
          \ once call connects."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.shared_short_code_controller"
...
securityDefinitions:
  auth:
    type: "basic"
    x-testValue: []
    x-skip-client-authentication: false
...