# rails5/app/controllers/recording_controller.rb
=begin
Ytel API V3

Ytel API version 3

OpenAPI spec version: 3.12

Generated by: https://github.com/swagger-api/swagger-codegen.git

=end
class RecordingController < ApplicationController

  def recording_deleterecording
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def recording_listrecording
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def recording_viewrecording
    # Your code here

    render json: {"message" => "yes, it worked"}
  end
end