# rails5/app/controllers/shared_short_code_controller.rb
=begin
Ytel API V3

Ytel API version 3

OpenAPI spec version: 3.12

Generated by: https://github.com/swagger-api/swagger-codegen.git

=end
class SharedShortCodeController < ApplicationController

  def keyword_lists
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def keyword_view
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def shortcode_getinboundsms
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def shortcode_listshortcode
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def shortcode_sendsms
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def shortcode_updateshortcode
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def shortcode_viewshortcode
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def template_lists
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def template_view
    # Your code here

    render json: {"message" => "yes, it worked"}
  end
end